Split Kokkos_Threads_Parallel files
