+++ core/src/Threads/Kokkos_ThreadsExec.hpp
@@ -26,8 +26,10 @@
 #include <impl/Kokkos_Spinwait.hpp>
 
 #include <Kokkos_Atomic.hpp>
+#include <Kokkos_Pair.hpp>
 
 #include <impl/Kokkos_ConcurrentBitset.hpp>
+#include <Threads/Kokkos_Threads.hpp>
 
 //----------------------------------------------------------------------------
